synx

+++ AnimationTest/动画效果和绘制测试/MyScrollViewController.m
@@ -39,7 +39,7 @@
     
     UITableViewCell *cell = [tableView dequeueReusableCellWithIdentifier:@"cell" forIndexPath:indexPath];
 
-    cell.textLabel.text = [NSString stringWithFormat:@"%ld",indexPath.row];
+    cell.textLabel.text = [NSString stringWithFormat:@"%ld",(long)indexPath.row];
 //    cell.contentView.subviews
     return cell;
 }
